Small cleanups in the test runner
For readability, change the data structure used to describe a failed
test from a simple list to a structure.

In the status reporter function, tab one further to make place for
"Skipped (irrelevant):", so that things are nicely aligned even when
the script is called with "--report-skipped-tests". Delete an
unreachable ECASE clause.

More OAOO-ness in UNEXPECTED-FAILURES.