Change the test runner to use randomly determined random values in the tests.
When the test runner is called, randomly determine a 256-bit random
seed and use this to seed a pseudo random number generator. Inside
WITH-TEST, set *RANDOM-STATE* to a fresh copy of this generator before
running the test. If any unexpected test results occur, report the
random seed in the final status report.

Add the option "--random-seed" to run-tests.sh to allow to force the
same random seed again in a later test run, change the option parser
to understand it and add a mechanism to report an option syntax error
from run-tests.lisp back to run-tests.sh.

Take measures to pass the random seed to the subprocesses created for
impure tests, in the process refactor the way to pass variables for more
OAOOness.