Faster ISQRT on small (about fixnum sized) numbers.
Make ISQRT use NUMBER-DISPATCH on FIXNUM and BIGNUM so that the compiler
can avoid most generic arithmetic in the fixnum part. This makes ISQRT
run about twice as fast on small inputs (fixnums and small bignums --
the latter as ISQRT always recurs into fixnum territory independent on
the input size).

The generated code is about 40 percent larger than the original.