Make the fixnum part of ISQRT recur without type dispatch.
Once the ISQRT recursion has reached a fixnum further recursive calls
are always fixnum sized, too. Implement the function so as to take
advantage of this fact. This gives an additional nice speedup: Now ISQRT
is about three times as fast on fixnums as the original version.

We can no longer use NUMBER-DISPATCH, so instead do an explicit
TYPECASE. The fixnum case needs two TRULY-THE type declarations to make
it compile into code without calls to generic arithmetic functions.