Small changes to improve readability.
Make the Macro explicitly take as an argument the variable to be used
to hold the expression's input.

Drop the "n-" prefix in the variables in the macro's body as this in
SBCL normally means "number of" and as the variable name being used by
the macro callers (which happens to be "n") is not visible here any
more.