First steps towards an ARX-based SXHASH.
The goal is to improve the distribution of SXHASH's values. This shall
be accomplished by using an ARX (add, rotate, xor) based mixing function
in SXHASH.

The mixing function is based on Tyche (see "Fast and Small Nonlinear
Pseudorandom Number Generators for Computer Simulation" by Samuel Neves
and Filipe Araujo), which itself is based on the ChaCha stream cipher.
Tyche uses 4 words of 32 bits as its state. I use 3 words of 64 bits to
get by with fewer processor registers. The rotation constants have been
chosen by experimentation and are subject to change.

This first commit introduces into x86-64 a word rotation VOP, the mixing
functions (initialization of the state and final mixing to return a
fixnum hash code), and uses them to replace SXHASH on fixnums.

I started with SXHASH on fixnums as it doesn't need to mix multiple
words of input, so can be made to work with little programming effort.
With regards to quality this is a bit unfair as the old fixnum hash
function does not even try to be nonlinear, so is very easy to beat.

The quality of the hash is such that INDEX-FOR-HASHING could be
trivialized to use only the least significant N bits of SXHASH's value.
Specifically, the amount of mixing done in this version makes SXHASH on
fixnums satisfy the strict avalanche criterion.

Regarding speed I have measured 15 clocks for the old implementation,
that is, a fixnum hash followed by INDEX-FOR-HASHING, and about 20
clocks for the ARX-based implementation, that is, a fixnum hash followed
by masking the least significant N bits according to some assumed
hash-table size.

I expect this implementation to be actually faster than the old one for
the more complex data types as it should be possible to hash additional
words in 3 clocks per word. It may even be suited for strings where I
expect it to be faster than the current implementation for strings of
length above about 5.